\name{scHSQ}
\alias{scHSQ}

\title{The scHSQ package}

\description{
The scHSQ package provides three primary functions. First, it allows for the identification of genes that fulfill the HSQ criteria using single cell RNA-Seq data. It also utilizes newly identified gene set to project cross-species data onto a single UMAP projection, allowing for the application of known cell type assignments across tissues that have undergone cell type-specific evolutionary divergence. Finally, it allows for visualization of a single gene's expression patterns on a this cross-species UMAP projection.
}
